pyproject.toml
[tool.poetry]
name = "instacrawl"
version = "0.1.0"
description = "A simple CLI Instagram crawler with a focus on algorithm analytics."
authors = [
  "Analetta Ehler <annie.ehler.4@gmail.com>"
]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.0.0"
torch = "^2.0.1"
rich = "^13.5.2"
face-recognition = "^1.3.0"
pynput = "^1.7.6"
opencv-python = "^4.8.0.74"
selenium = "^4.11.2"
selenium-wire = "^5.1.0"
beautifulsoup4 = "^4.12.2"
dtale = "^3.3.0"
eva-decord = "^0.6.1"
kaleido = "0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
